# openapi_v1.0.10.yaml
openapi: 3.0.0
info:
  title: PMS APIs
  description: >
    This API requires OAuth 2.0 Client Credentials flow for authentication.
    The client must pass the `client_id` and `client_secret` as a Base64-encoded string in the Authorization header.
    The format should be `Basic <Base64(client_id:client_secret)>`.
  version: 1.0.10
servers:
  - url: https://devapi.10bridge.io/
paths:
...
  /Opargo/dev/practitioners:
    post:
      operationId: Providers
      tags:
        - GetProviders
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMetaData"

      responses:
        "200":
          description: Get practice Providers for a given practice.
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Practitioner"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
  /Opargo/dev/referring_providers:
    post:
      operationId: ReferringProviders
      tags:
        - GetReferringProviders
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMetaData"

      responses:
        "200":
          description: Get refferring providers for a given practice.
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Practitioner"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
  /Opargo/dev/patients/search:
    post:
      operationId: Patients
      tags:
        - SearchPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRequest"

      responses:
        "200":
          description: Patients found - search patients
          content:
            application/json:
              schema:
                type: object
                properties:
                  patients:
                    type: array
                    items:
                      $ref: "#/components/schemas/Patient"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "404":
          description: No patients Found - The search returned no results.
        "500":
          description: Internal Server Error - An unexpected error occurred.
...
  /Opargo/dev/schedule/search:
    post:
      operationId: Schedule
      tags:
        - GetProviderSchedule
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRequest"

      responses:
        "200":
          description: search Schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedule:
                    type: array
                    items:
                      $ref: "#/components/schemas/Schedule"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
...
  /Opargo/dev/appointment_notes:
    post:
      operationId: AppointmentNotes
      tags:
        - GetNotes
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentNotesRequest"

      responses:
        "200":
          description: get Appointment notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppointmentNotes"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
  /Opargo/dev/appointments/search:
    post:
      operationId: Appointments
      tags:
        - GetAppontments
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentSearchRequest"

      responses:
        "200":
          description: search Appointments
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Appointment"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "404":
          description: No Appointments Found - The search returned no results.
        "500":
          description: Internal Server Error - An unexpected error occurred.
  /Opargo/dev/appointments:
    put:
      operationId: CancelAppointment
      tags:
        - CancelAppointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelAppointmentRequest"

      responses:
        "200":
          description: Appointment cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
    post:
      operationId: BookAppointment
      tags:
        - BookAppointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewAppointmentRequest"
      responses:
        "200":
          description: Appointment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
  /Opargo/dev/patients:
    post:
      operationId: Patient
      tags:
        - CreatePatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientCreateRequest"

      responses:
        "200":
          description: patinet created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Bad Request - The request is invalid or missing required parameters.
        "401":
          description: Unauthorized - Invalid or expired token
        "500":
          description: Internal Server Error - An unexpected error occurred.
...
security:
  - OAuth2: []
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://devapi.10bridge.io/api/auth/v2/token
          scopes: {}
  schemas:
...
    Practitioner:
      type: object
      properties:
        Practitioner_id:
          type: string
        speciality:
          type: string
        is_physician:
          type: boolean
        is_practice_member:
          type: boolean
        full_name:
          type: string
        abbreviation:
          type: string
        is_active:
          type: string
      required:
        - Practitioner_id
        - speciality
        - full_name
        - abbreviation
...
    Patient:
      allOf:
        - $ref: "#/components/schemas/NewPatient"
      type: object
      properties:
        patientProfileID:
          type: string
        patient_id:  
          type: string
        practitioner_id:
          type: string
        insurances:
          type: array
          items:
            $ref: "#/components/schemas/InsurancePolicy"
      required:
        - profile_id
        - practitioner_id
...
    Schedule:
      type: object
      properties:
        scheduled_id:
          type: string
          description: "appointment id "
        past_appointment:
          type: boolean
        status:
          type: string
          description: "either it is bumped or scheduled "
        service_category:
          type: string
          description: "Allied Health or Child Care or Dental etc Ref: https://build.fhir.org/appointment-definitions.html#Appointment.serviceCategory"
        service_type:
          type: string
          description: "The type of appointments that can be booked into this slot (ideally this would be an identifiable service - which is at a location, rather than the location itself). If provided then this overrides the value provided on the Schedule resource"
        name:
          type: string
          description: "This MAY be used to describe what the schedule is for"
        comments:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        provider:
          $ref: "#/components/schemas/Practitioner"
        location:
          $ref: "#/components/schemas/Location"
      required:
        - scheduled_id
        - start_date
        - end_date
        - service_type
        - provider
        - location
        - past_appointment
        - status
...
    RequestMetaData:
      type: object
      properties:
        siteID:
          type: string
        customerName:
          type: string
      required: # Added the required fields here
        - siteID
        - customerName
...
    ScheduleRequest:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/RequestMetaData"
        data:
          type: object
          properties:
            appointment_type:
              type: string
            patientProfileId:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            start_date:
              type: string
            end_date:
              type: string
          required:
            - patientProfileId
            - first_name
            - last_name
            - start_date
            - end_date
    AppointmentSearchRequest:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/RequestMetaData"
        data:
          type: object
          properties:
            patient_id:
              type: string
            practioner_abbreviation:
              type: string
            start_date:
              type: string
              format: "date"
            end_date:
              type: string
              format: "date"
    PatientRequest:
      type: object
      properties:
        meta: 
          $ref: '#/components/schemas/RequestMetaData'
        data:
          type: object
          properties:
            first_name:
              type: string
            last_name: 
              type: string
            date_of_birth:
              type: string 
          required:
            - last_name
            - date_of_birth          
    AppointmentNotesRequest:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/RequestMetaData"
        data:
          type: object
          properties:
            appointment_id:
              type: string
            patient_profile_id:
              type: string
    AppointmentNotes:
      type: object
      properties:
        note:
          type: string
        appointment:
          type: string
        patient_profile_id:
          type: string
...